Camera is now in the middle of screen. Made movement follow float values to not be restricted by grid.

+++ Assets/Scripts/Enemy.cs
@@ -19,7 +19,7 @@ public class Enemy : MovingObject {
         base.Start();
 	}
 
-    protected override void AttemptMove<T>(int xDir, int yDir) {
+    protected override void AttemptMove<T>(float xDir, float yDir) {
         skipMove = !skipMove;
         if (skipMove)
             return;
